refactor: rename `install` command to `add`
The word `add` more accurately reflects what is happening here, since the command can add any number
of services using the `--count` argument, and it is also intended to be used multiple times; you
don't really run an installation more than once.
