+++ sn-node-manager/src/add_service.rs
